Refactor Plex OAuth integration: streamline routes, enhance token handling, and update server retrieval to use JSON response.
