+++ bazarr/plex/service.py
@@ -3,13 +3,38 @@ import os
 import requests
 from app.config import settings, save_settings
 from flask import session
-from .models import PlexServerConnection, PlexServerToken
 
 PLEX_CLIENT_ID = os.getenv('PLEX_CLIENT_ID', 'bazarr')
 PLEX_CLIENT_SECRET = os.getenv('PLEX_CLIENT_SECRET', '')
 PLEX_REDIRECT_URI = os.getenv('PLEX_REDIRECT_URI', 'http://localhost:6767/api/plex/oauth/callback')
 PLEX_OAUTH_URL = 'https://plex.tv/users/sign_in.json'
-PLEX_RESOURCES_URL = 'https://plex.tv/api/resources?includeHttps=1'
+PLEX_RESOURCES_URL = 'https://plex.tv/api/v2/resources?includeHttps=1'
+
+class PlexServerConnection:
+    def __init__(self, name, address, port, uri, local, ssl):
+        self.name = name
+        self.address = address
+        self.port = port
+        self.uri = uri
+        self.local = local
+        self.ssl = ssl
+
+    def to_dict(self):
+        return {
+            'name': self.name,
+            'address': self.address,
+            'port': self.port,
+            'uri': self.uri,
+            'local': self.local,
+            'ssl': self.ssl,
+        }
+
+class PlexServerToken:
+    def __init__(self, token):
+        self.token = token
+
+    def to_dict(self):
+        return {'token': self.token}
 
 class PlexService:
     @staticmethod
@@ -51,24 +76,29 @@ class PlexService:
             return None
         resp = requests.get(PLEX_RESOURCES_URL, headers={
             'X-Plex-Token': token,
-            'Accept': 'application/xml',
+            'Accept': 'application/json',
         })
         if resp.status_code != 200:
             return None
-        import xml.etree.ElementTree as ET
-        root = ET.fromstring(resp.text)
-        servers = []
-        for device in root.findall('.//Device'):
-            for conn in device.findall('Connection'):
-                servers.append(PlexServerConnection(
-                    name=device.get('name'),
-                    address=conn.get('address'),
-                    port=conn.get('port'),
-                    uri=conn.get('uri'),
-                    local=conn.get('local'),
-                    ssl=conn.get('protocol') == 'https',
-                ))
-        return servers
+        # Parse JSON response
+        try:
+            resources = resp.json()
+            servers = []
+            for resource in resources:
+                if resource.get('provides') == 'server' and resource.get('owned', False):
+                    connections = resource.get('connections', [])
+                    for conn in connections:
+                        servers.append(PlexServerConnection(
+                            name=resource.get('name'),
+                            address=conn.get('address'),
+                            port=str(conn.get('port', 32400)),
+                            uri=conn.get('uri'),
+                            local=str(conn.get('local', False)).lower(),
+                            ssl=conn.get('protocol') == 'https',
+                        ))
+            return servers
+        except (ValueError, KeyError) as e:
+            return None
 
     @staticmethod
     def save_selected_server(data):
